pyproject.toml
[tool.poetry]
name = "kahvi"
version = "0.1.0"
description = ""
authors = ["Karri Lumivirta <lumivirta.karri@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10, <3.13"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flask-sqlalchemy = "^3.1.1"
robotframework-seleniumlibrary = "^6.6.1"
requests = "^2.32.3"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.7"
robotframework = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
